+++ lib/src/theme/component/constrained_screen.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+import '../../theme_service.dart';
+import '../res/layout.dart';
+
+class ConstrainedScreen extends StatelessWidget {
+  const ConstrainedScreen({
+    super.key,
+    required this.child,
+  });
+
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      alignment: Alignment.center,
+      color: context.color.surface,
+      child: ConstrainedBox(
+        constraints: const BoxConstraints(
+          maxWidth: Breakpoints.desktop,
+        ),
+        child: child,
+      ),
+    );
+  }
+}