+++ lib/src/theme/component/pop_button.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+import '../../theme_service.dart';
+import 'button/button.dart';
+
+class PopButton extends StatelessWidget {
+  const PopButton({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Button(
+      icon: 'arrow-left',
+      color: context.color.text,
+      type: ButtonType.flat,
+      onPressed: () => Navigator.pop(context),
+    );
+  }
+}